some structural code refactorment to sate machine of Hadith,
still have many false possitives due to runon words

+++ trunk/ATSarf/src/common/common.h
@@ -89,7 +89,7 @@ public:
 typedef struct text_info_
 {
 	QString *text;
-	long long start,finish;
+	long start,finish;
 }text_info;
 
 class multiply_params
